refactored some of auth.py functionality into db.py
